refactor: apllying mvc pattern as good practice
