+++ my-turno-web/src/app/components/TextInput/index.tsx
@@ -1,23 +1,7 @@
 "use client";
 
 import EyeIcon from "@component/app/assets/icons/Eye";
-import { FieldError, UseFormRegisterReturn } from "react-hook-form";
-
-export interface TextInputProps {
-  id: string;
-  placeholder?: string;
-  type?: InputType;
-  label?: string;
-  errors?: FieldError;
-  registerOptions?: UseFormRegisterReturn;
-  hideErrorMessage?: boolean;
-}
-
-export enum InputType {
-  text = "text",
-  password = "password",
-  email = "email",
-}
+import { InputType, TextInputProps } from "./types";
 
 export default function TextInput(props: TextInputProps) {
   const {
