+++ my-turno-web/src/app/pages/signup/page.tsx
@@ -1,4 +1,3 @@
-import Navbar from "../../components/Navbar";
 import Signup from "../../sections/Signup";
 
 export default function SignupSection() {
