+++ my-turno-web/src/app/sections/Reservations/types.ts
@@ -1,0 +1,4 @@
+import { ReservationFormData } from "@component/schema";
+import { z } from "zod";
+
+export type ReservationForm = z.infer<typeof ReservationFormData>;
